mds: use inode_t::layout for dir layout policy
Remove the file_layout_policy_t struct, which was just
a ceph_file_layout, and store it in the inode_t.  Rip out
all the annoying code that put this on the heap.

This is currently *only* compile-tested. Also, an earlier
version of the patch revved the CEPH_MDS_PROTOCOL value, but this
one doesn't. That might or might not be necessary to add back in?

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>
Signed-off-by: Greg Farnum <greg@inktank.com>